+++ src/bitmap_displayer_mac.c
@@ -13,7 +13,7 @@ void cursor_home() {
 } //Clears the terminal output (SLOW)
 
 
-char *map_pixel_value_to_char(uint8_t character_entry) {
+char *map_pixel_value_to_char(uint16_t character_entry) {
   if (character_entry == 0) {
     return BLOCK_EMPTY;
   }
@@ -38,15 +38,15 @@ char *map_pixel_value_to_char(uint8_t character_entry) {
 }
 
 
-int print_bitmap(uint8_t *bitmap_loc) {
+int print_bitmap(uint16_t *bitmap_loc) {
   // Prints the current bitmap out to the screen.
   int i = 0;
   int col_number = 0;
   
   while (i < SCREEN_HEIGHT * SCREEN_WIDTH) {
   col_number = i % SCREEN_WIDTH;
-    uint8_t *current_position = bitmap_loc + i;
-    uint8_t current_pixel_value = *current_position;
+    uint16_t *current_position = bitmap_loc + i;
+    uint16_t current_pixel_value = *current_position;
     char * current_char = map_pixel_value_to_char(current_pixel_value);
     printf("%s", current_char);
     i++;
